Fix bug with ints being created instead of matching the input dtype

+++ fast_plotter/plotting.py
@@ -238,7 +238,7 @@ def add_missing_vals(x, expected_xs, y_values=[], fill_val=0):
     insert = np.isin(expected_xs, x)
     new_ys = []
     for y in y_values:
-        new = np.full_like(expected_xs, fill_val)
+        new = np.full_like(expected_xs, fill_val, dtype=y.dtype)
         new[insert] = y
         new_ys.append(new)
     return expected_xs[:], new_ys
